classification model
The original dataset was too small, leading to an overfitted model and class imbalance, so the dataset was replaced. This process includes: preparing the new dataset with labeled data, building the model and saving it as a .pkl file, and finally using joblib to load the .pkl file and directly perform predictions with the model.